Remove dead code and organise runtime files into core/, content/, popup/
Dead code removed:
- offscreen.html + offscreen.js: orphaned MV3 offscreen-document ZIP
  pipeline. Nothing calls chrome.offscreen (the manifest never had the
  'offscreen' permission), nothing sends its 'startZip' message and
  nothing listens for its offscreenProgress/Done/Error replies — the
  service worker builds ZIPs itself via importScripts(JSZip).
- scripts/manual-fallback.js: console-paste workaround for the
  scrambled-pages bug (issue #2), obsolete since the proper fix landed
  (c24b798); the README bootloader it depended on is long gone.
- manifest web_accessible_resources: lib/jszip.min.js is loaded via
  importScripts (Chrome) / background.scripts (Firefox), never fetched
  by a page through runtime.getURL — exposing it only added an
  extension-fingerprinting surface.
- assets/.gitkeep: the folder ships three screenshots.

Reorganisation (same filenames, new folders):
- core/     shared modules loaded in multiple contexts: settings.js,
            cdl-features-core.js, cdl-home-core.js, cdl-badge-core.js,
            cdl-comicinfo.js
- content/  content scripts: content_*.js, cdl-embed-settings.js, and
            extract-bridge.js (moved out of scripts/, which now holds
            dev tooling only)
- popup/    popup.html + popup.js

All references updated: manifest content_scripts/default_popup,
background.js importScripts, legacy/options.html script tag,
build-release.ps1 runtimePaths + generated Firefox background.scripts,
bump-version.ps1, and test require/read paths + manifest assertions.

Verified: all 6 test suites pass (433 checks), every path referenced by
manifest.json and build-release.ps1 exists, and web-ext lint reports
0 errors / 0 warnings on a staged copy of the Firefox package layout.
