+++ content/cdl-embed-settings.js
