+++ content/extract-bridge.js
