+++ core/cdl-badge-core.js
