+++ core/cdl-comicinfo.js
