+++ legacy/options.html
@@ -51,7 +51,7 @@
   <div class="toast-wrap" id="toast-wrap" aria-live="polite"></div>
   <input type="file" id="import-file" accept="application/json,.json" hidden />
 
-  <script src="../settings.js"></script>
+  <script src="../core/settings.js"></script>
   <script src="options.js"></script>
 </body>
 </html>
