--- offscreen.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-<head>
-  <meta charset="UTF-8">
-  <title>Comix Downloader — Offscreen</title>
-</head>
-<body>
-  <!-- Document offscreen invisible utilisé pour créer des Blob URL et déclencher les téléchargements -->
-  <script src="lib/jszip.min.js"></script>
-  <script src="offscreen.js"></script>
-</body>
-</html>
