+++ scripts/build-release.ps1
@@ -34,23 +34,21 @@ function Reset-Directory([string]$Path) {
 function Copy-ReleaseFiles([string]$Destination) {
   $runtimePaths = @(
     "background.js",
-    "content_title.js",
-    "content_notices.js",
-    "cdl-features-core.js",
-    "cdl-home-core.js",
-    "cdl-badge-core.js",
-    "cdl-comicinfo.js",
-    "content_features.js",
-    "content_home.js",
-    "content_profile.js",
-    "cdl-embed-settings.js",
-    "scripts/extract-bridge.js",
+    "content/content_title.js",
+    "content/content_notices.js",
+    "content/content_features.js",
+    "content/content_home.js",
+    "content/content_profile.js",
+    "content/cdl-embed-settings.js",
+    "content/extract-bridge.js",
+    "core/settings.js",
+    "core/cdl-features-core.js",
+    "core/cdl-home-core.js",
+    "core/cdl-badge-core.js",
+    "core/cdl-comicinfo.js",
     "manifest.json",
-    "offscreen.html",
-    "offscreen.js",
-    "popup.html",
-    "popup.js",
-    "settings.js",
+    "popup/popup.html",
+    "popup/popup.js",
     "legacy/options.html",
     "legacy/options.js",
     "legacy/options.css",
@@ -120,9 +118,9 @@ $firefoxManifest = $manifest | ConvertTo-Json -Depth 20 | ConvertFrom-Json
 $firefoxManifest.background = [ordered]@{
   scripts = @(
     "lib/jszip.min.js",
-    "settings.js",
-    "cdl-features-core.js",
-    "cdl-comicinfo.js",
+    "core/settings.js",
+    "core/cdl-features-core.js",
+    "core/cdl-comicinfo.js",
     "background.js"
   )
 }
